Move resolveProjectRef to a common library
